some added new features

+++ src/com/company/Player_die/Player.java
@@ -6,7 +6,7 @@ public class Player {
 
     //Property of our Player
     private String name;
-    private double money;
+    private int money;
     private int doubleCounter;
     private boolean isBankrupt;
     private int turn;
@@ -51,11 +51,11 @@ public class Player {
         this.name = name;
     }
 
-    public double getMoney() {
+    public int getMoney() {
         return money;
     }
 
-    public void setMoney(double money) {
+    public void setMoney(int money) {
         this.money = money;
     }
